pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "oncorounds"
version = "1.0.0"
description = "OncoRounds: Interactive benchmark for LLM clinical reasoning in oncology"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Keno Bressem", email = "keno[dot]bressem[at]tum[dot]de"}
]
keywords = ["oncology", "benchmark", "llm", "medical", "clinical-reasoning"]
dependencies = [
    "jsonschema>=4.23.0",
    "openai>=1.40.0",
    "python-dotenv>=1.0.1",
    "tqdm>=4.66.0",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

[project.urls]
Homepage = "https://github.com/kbressem/oncorounds"
Documentation = "https://kbressem.github.io/oncorounds/"
Repository = "https://github.com/kbressem/oncorounds"
Issues = "https://github.com/kbressem/oncorounds/issues"
Changelog = "https://github.com/kbressem/oncorounds/blob/main/CHANGELOG.md"

[project.scripts]
oncorounds-score = "oncorounds.scoring:_cli"

[project.optional-dependencies]
dev = [
    "mypy>=1.10.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
addopts = "--maxfail=1 --disable-warnings"
testpaths = ["tests"]

[tool.coverage.run]
source = ["oncorounds"]
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
